backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
resend==0.7.0

# Payments
stripe==7.8.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Formula Engine
simpleeval==0.9.13

# Monitoring (optional)
sentry-sdk[fastapi]==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
